feat: allow to enable/disable some plugins
List of plugins for gateway (apisix container) is currently configured
statically in the ConfigMap and there is no way to customize the list.
That means that e.g. `zipkin` plugin for observability is always enabled,
but `opentelemetry` is not and it is not possible to enabled it (easily).

This just a proposal, how it could be done in one way.

I was also thinking about:

```
gateway:
  pluginsOverride:
    - plugin1
    - plugin2
    - ...
```

Or instead of adjusting ConfigMap template to allow for swappping it
with `existingConfigMap`, e.g.:

```
gateway:
  existingConfigMap: "my-apisix-configmap"
```

Which would cause to not render `apisix-configmap.yaml` at all and
instead switch `configuration` volume to provided custom ConfigMap.

Feedback is welcome.